Add statuscmd patch
This patch allows clicking on statusbar "blocks". These click events
trigger a signal on the corresponding dwmblock defined in the dwmblocks
repo. I have not yet uploaded the dwmblocks repository anywhere. I'm not
sure if it'll be here or where, but I will. Anyway, this patch allows
for clickable blocks which can be handled by any program. Currently I'm
only using sh scripts. Those can be found in my dotfiles