Trust NET_WM_FOCUS and apply centeredmaster patch
Actually I kind of hated that it doesn't respect actions, making my
notifications kind of useless (it's 2021 anyway), but at the same time I
can understand why this wasn't built in to dwm.

About centeredmaster, I think it works well in some cases, where most of
the time there's nothing relevant at all in the center of the screen,
making me tilt my head.

This commit also binds super+v to a personal script I have and don't
know where to put it because dwm seems to handle it way better than
sxhkd, which is a real bummer.