Move shortcuts away from dwm
The idea is to move as much as I can away from dwm to avoid recompiling
each time I decide to change a shortcut. Rationale is that sxhkd is way
better at doing this, even if it's performance is a little subpar. I'm
actually thinking of patching dwm with dwmc and anybar and move the
config completely outside this.